# RadiKeep/openapi/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "RadiKeep | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://127.0.0.1:7262/"
    }
  ],
  "paths": {
...
    "/api/notifications": {
      "get": {
        "tags": [
          "NotificationsApi"
        ],
        "summary": "お知らせ一覧を取得する",
        "operationId": "ApiNotificationList",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfNotificationListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfEmptyData"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "required": [
          "code",
          "message"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": [
              "null",
              "string"
            ]
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "ApiResponseOfEmptyData": {
        "required": [
          "success",
          "data",
          "error",
          "message"
        ],
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/EmptyData"
              }
            ]
          },
          "error": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ApiError"
              }
            ]
          },
          "message": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ApiResponseOfNotificationListResponse": {
        "required": [
          "success",
          "data",
          "error",
          "message"
        ],
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/NotificationListResponse"
              }
            ]
          },
          "error": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ApiError"
              }
            ]
          },
          "message": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "EmptyData": {
        "type": "object"
      },
...
      "NoticeCategory": {
        "type": "integer"
      },
      "NotificationEntry": {
        "type": "object",
        "properties": {
          "logLevel": {
            "type": "string"
          },
          "category": {
            "$ref": "#/components/schemas/NoticeCategory"
          },
          "message": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "isRead": {
            "type": "boolean"
          }
        }
      },
...
      "NotificationListResponse": {
        "required": [
          "totalRecords",
          "page",
          "pageSize",
          "recordings"
        ],
        "type": "object",
        "properties": {
          "totalRecords": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "page": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "pageSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "recordings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NotificationEntry"
            }
          }
        }
      },
...
    }
  },
...
}